Resolved warnings when running make

+++ source/test_serialization.c
@@ -56,8 +56,8 @@ int testSerialize(){
         | nkeys | key1   | key2   | key3   |
 	*/
 
-	assert(size == (sizeof(int) + strlen(key1)+1 + strlen(key2)+1 + strlen(key3)+1));
-	result = result && (size == (sizeof(int) + strlen(key1)+1 + strlen(key2)+1 + strlen(key3)+1));
+	assert((uint32_t) size == (sizeof(int) + strlen(key1)+1 + strlen(key2)+1 + strlen(key3)+1));
+	result = result && ((uint32_t) size == (sizeof(int) + strlen(key1)+1 + strlen(key2)+1 + strlen(key3)+1));
 
 	assert(ntohl(*((int*)keys_buf)) == 3);
     	result = result && (ntohl(*((int*)keys_buf)) == 3);
